backend/requirements.txt
langgraph
langchain
langchain-groq
groq
python-dotenv
chromadb 
langchain-community 
langchain-openai
requests
trafilatura
beautifulsoup4

